requirements.txt
pip==24.2
Django==5.0.7
djangorestframework==3.15.2
psycopg2-binary==2.9.9
python-dotenv==1.0.1
pillow==10.4.0
django-filter==24.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
setuptools==72.1.0
stripe==10.6.0
requests==2.32.3
forex-python==1.8
